+++ zoo-modules/searchable-select-module/SearchableSelect.svelte
@@ -14,6 +14,11 @@
 			<zoo-preloader></zoo-preloader>
 		{/if}
 		<slot bind:this={_selectSlot} name="selectelement"></slot>
+		{#if _valueSelected}
+			<div class="close" on:click="{e => handleCrossClick()}">
+				<svg width="18" height="18" viewBox="0 0 24 24"><path d="M19 6.41L17.59 5 12 10.59 6.41 5 5 6.41 10.59 12 5 17.59 6.41 19 12 13.41 17.59 19 19 17.59 13.41 12z"/></svg>
+			</div>
+		{/if}
 	{:else}
 		<zoo-select labelposition="{labelposition}" linktext="{linktext}" linkhref="{linkhref}" linktarget="{linktarget}"
 			labeltext="{labeltext}" inputerrormsg="{inputerrormsg}" infotext="{infotext}" valid="{valid}">
@@ -24,6 +29,14 @@
 
 <style type='text/scss'>
 	@import "variables";
+
+	.close {
+		display: inline-block;
+		position: absolute;
+		top: 39%;
+		right: 4%;
+		cursor: pointer;
+	}
 
 	:host {
 		position: relative;
@@ -105,6 +118,7 @@
 	let _prevValid;
 	let options;
 	let _isMobile;
+	let _valueSelected;
 	let tooltipText;
 
 	beforeUpdate(() => {
@@ -138,6 +152,7 @@
 			}
 
 			_selectElement.classList.add('searchable-zoo-select');
+			_selectElement.addEventListener('change', e => _valueSelected = e.target.value ? true : false);
 			_hideSelectOptions();
 			changeValidState(valid);
 	    });
@@ -224,5 +239,11 @@
 	const isMobile = () => {
 		const index = navigator.appVersion.indexOf("Mobile");
 		return (index > -1);
+	}
+
+	const handleCrossClick = () => {
+		_selectElement.value = null;
+		_selectElement.dispatchEvent(new Event("change"));
+		handleOptionClick();
 	}
 </script>