+++ zoo-modules/select-module/Select.svelte
@@ -11,6 +11,11 @@
 			{#if loading}
 				<zoo-preloader></zoo-preloader>
 			{/if}
+			{#if _valueSelected}
+				<div class="close" on:click="{e => handleCrossClick()}">
+					<svg width="18" height="18" viewBox="0 0 24 24"><path d="M19 6.41L17.59 5 12 10.59 6.41 5 5 6.41 10.59 12 5 17.59 6.41 19 12 13.41 17.59 19 19 17.59 13.41 12z"/></svg>
+				</div>
+			{/if}
 		{/if}
 	</span>
 	<zoo-input-info class="input-info" valid="{valid}" inputerrormsg="{inputerrormsg}" infotext="{infotext}">
@@ -20,6 +25,14 @@
 <style type='text/scss'>
 	@import "variables";
 	@import "input";
+
+	.close {
+		display: inline-block;
+		position: absolute;
+		top: 33%;
+		right: 7%;
+		cursor: pointer;
+	}
 
 	.arrows {
 		position: absolute;
@@ -95,6 +108,7 @@
 	let _multiple = false;
 	let _slottedSelect;
 	let _selectSlot;
+	let _valueSelected;
 
 	beforeUpdate(() => {
 		if (valid != _prevValid) {
@@ -110,6 +124,7 @@
 			if (select.multiple === true) {
 				_multiple = true;
 			}
+			_slottedSelect.addEventListener('change', e => _valueSelected = e.target.value ? true : false);
 			changeValidState(valid);
 	    });
 	});
@@ -122,5 +137,10 @@
 				_slottedSelect.classList.remove('error');
 			}
 		}
+	}
+
+	const handleCrossClick = () => {
+		_slottedSelect.value = null;
+		_slottedSelect.dispatchEvent(new Event("change"));
 	}
 </script>